# Conflicts:
#	app/src/main/java/com/app/ecarepro/ui/MainActivity.kt

+++ app/src/main/java/com/app/ecarepro/ui/leave/apply_leave/ApplyLeaveViewModel.kt
@@ -37,7 +37,7 @@ class ApplyLeaveViewModel @Inject constructor(
     fun leaveApply(leaveID: Int,
                    fromDate: String,
                    tillDate:String,
-                   duration:Int,
+                   duration:Double,
                    halfdayDTL: List<HalfdayDTL>,
                    reason:String,
                    attachment:String,
