+++ app/src/main/java/com/app/ecarepro/ui/leave/staff/StaffApplyLeaveViewModel.kt
@@ -30,7 +30,7 @@ class StaffApplyLeaveViewModel @Inject constructor(
     fun leaveApply(leaveID: Int,
                    fromDate: String,
                    tillDate:String,
-                   duration:Int,
+                   duration:Double,
                    halfdayDTL: List<HalfdayDTL>,
                    reason:String,
                    attachment:String,
